Move GTFSRealtime class to gtfs_rt_detail
